requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
tqdm
wandb
segmentation_models_pytorch
albumentations